# c01/backend/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "African Impact Challenge",
      "description": "",
      "version": "1.0"
    },
    "basePath":"/api",
    "produces": ["application/json"],
    "paths": {
...
      "/company/change_company_password": {
        "put": {
          "x-swagger-router-controller": "home",
          "operationId": "1",
          "tags": ["/api/company"],
          "description": "Changing the company's password.",
          "parameters": [
            {"name":"authentication-token","in":"header","required":true},
            {"name":"request body","in":"body","schema":{
              "type":"object",
              "properties":{
                "newPassword":{
                  "type":"string",
                  "required":true
                }
              }
            }}
          ],
          "responses": {}
        }
      },
      "/company/change_company_info": {
        "put": {
          "x-swagger-router-controller": "bar",
          "operationId": "2",
          "tags": ["/api/company"],
          "description": "Changing company's infomation.",
          "parameters": [
            {"name":"authentication-token","in":"header","required":true},
            {"name":"request body","in":"body","schema":{
              "type":"object",
              "properties":{
                "name":{
                  "type":"string",
                  "required":true
                },
                "email":{
                  "type":"string",
                  "required":true
                },
                "location":{
                  "type":"string"
                },
                "industry":{
                  "type":"string"
                },
                "website":{
                  "type":"string"
                },
                "description":{
                  "type":"string"
                }
              }
            }}
          ],
          "responses": {}
        }
      },
      "/company/partner_view_id/{company_id}": {
        "get": {
          "x-swagger-router-controller": "bar",
          "operationId": "sprint3-1",
          "tags": ["/api/company"],
          "description": "Partner could get the company's more information by its ID.",
          "parameters": [
            {"name":"company_id","in":"path","required":true}
          ],
          "responses": {}
        }
      },
      "/company/partner_view_name/{company_name}": {
        "get": {
          "x-swagger-router-controller": "bar",
          "operationId": "sprint3-2",
          "tags": ["/api/company"],
          "description": "Partner could get the company's more information by its name.",
          "parameters": [
            {"name":"company_name","in":"path","required":true}
          ],
          "responses": {}
        }
      },
      "/company/show_company_info_email/{company_email}": {
        "get": {
          "x-swagger-router-controller": "bar",
          "operationId": "sprint3-3",
          "tags": ["/api/company"],
          "description": "Get the company's information by its email.",
          "parameters": [
            {"name":"company_email","in":"path","required":true}
          ],
          "responses": {}
        }
      },
      "/company/show_company_info_id/{company_id}": {
        "get": {
          "x-swagger-router-controller": "bar",
          "operationId": "3",
          "tags": ["/api/company"],
          "description": "Get the company's information by its ID.",
          "parameters": [
            {"name":"company_id","in":"path","required":true}
          ],
          "responses": {}
        }
      },
      "/company/show_company_info_name/{company_name}": {
        "get": {
          "x-swagger-router-controller": "bar",
          "operationId": "4",
          "tags": ["/api/company"],
          "description": "Get the company's information by its name",
          "parameters": [
            {"name":"company_name","in":"path","required":true}
          ],
          "responses": {}
        }
      },
...
      "/company/login/company_email": {
        "post": {
          "x-swagger-router-controller": "bar",
          "operationId": "6",
          "tags": ["/api/company"],
          "description": "Login to company account by email",
          "parameters": [
            {"name":"request body","in":"body","schema":{
              "type":"object",
              "properties":{
                "email":{
                  "type":"string",
                  "required":true
                },
                "password":{
                  "type":"string",
                  "required":true
                }}
              }
            }            
          ],
          "responses": {}
        }
      },
      "/company/login/company_name": {
        "post": {
          "x-swagger-router-controller": "bar",
          "operationId": "7",
          "tags": ["/api/company"],
          "description": "Login to company account by name",
          "parameters": [
            {"name":"request body","in":"body","schema":{
              "type":"object",
              "properties":{
                "name":{
                  "type":"string",
                  "required":true
                },
                "password":{
                  "type":"string",
                  "required":true
                }
              }}
            }            
          ],
          "responses": {}
        }
      },
      "/company/company_register": {
        "post": {
          "x-swagger-router-controller": "bar",
          "operationId": "8",
          "tags": ["/api/company"],
          "description": "Company registration",
          "parameters":[
            {"name":"request body","in":"body","schema":{
              "type":"object",
              "properties":{
                "name":{
                  "type":"string",
                  "required":true
                },
                "email":{
                  "type":"string",
                  "required":true
                },
                "password":{
                  "type":"string",
                  "required":true
                },
                "location":{
                  "type":"string"
                },
                "industry":{
                  "type":"string"
                },
                "website":{
                  "type":"string"
                },
                "description":{
                  "type":"string"
                }
              }}
            } 
          ],
          "responses": {}
        }
      },
...
  }
}